Adds two internal utility functions:
1. `is_factor_term` and
2. `term_variables`.

`is_factor_term()` will check if a specified model term is a factor
or not. This is useful for terms in models like `f1:f2` or `f1:x1`
where `f1` and `f2` are factors and `x1` is numeric. This function
returns `TRUE` for terms involving only factors, e.g. `f1:f2`. For
terms involving numeric variables it returns `FALSE`

`term_variables()` will identify the names of variables, involved in
a model term (in the sense of `attr(terms(model) "term.labels")`)